python/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
PyMySQL==1.1.0
pandas==2.1.1
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
